refactor(native): isolate application host tests
